+++ AvalancheGamesWeb/Controllers/ScoreController.cs
@@ -125,6 +125,10 @@ namespace AvalancheGamesWeb.Controllers
         {
             try
             {
+                if (!ModelState.IsValid)
+                {
+                    return View(collection);
+                }
                 // TODO: Add insert logic here
                 using (ContextBLL ctx = new ContextBLL())
                 {
@@ -170,6 +174,10 @@ namespace AvalancheGamesWeb.Controllers
         {
             try
             {
+                if (!ModelState.IsValid)
+                {
+                    return View(collection);
+                }
                 // TODO: Add update logic here
                 using (ContextBLL ctx = new ContextBLL())
                 {
@@ -212,6 +220,10 @@ namespace AvalancheGamesWeb.Controllers
         {
             try
             {
+                if (!ModelState.IsValid)
+                {
+                    return View(collection);
+                }
                 using (ContextBLL ctx = new ContextBLL())
                 {
                     ctx.DeleteScore(id);
